feat: Extractor Arbitrum - getFilterChanges
